fashion_store/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
mysqlclient==2.2.0
Pillow==10.0.1
python-decouple==3.8